Single .desktop file
qTsConverter.desktop was used for manual installation (universal package
or make install), org.guerinoni.desktop for flatpak but with normal
installation both had do be installed. This created a duplicate in
desktop environments menus.  Keep in mind that the remaining .desktop
file doesn't have an absolute path in the Exec field, so /usr/local/bin
must be in $PATH when installing manually.

Directories end with / in cmake install